Refactor: Move integration tests and add schema validation tests
